restructure to synchronize with enact
