--- lib/src/color_tokens/color_tokens.dart
@@ -1,7 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright 2025 LG Electronics Inc.
- * SPDX-License-Identifier: LicenseRef-LGE-Proprietary
- */
-
-export 'semantic/color_semantic.dart';
-export 'primitive/color_primitive.dart';
