--- lib/src/font_tokens/font_tokens.dart
@@ -1,7 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright 2025 LG Electronics Inc.
- * SPDX-License-Identifier: LicenseRef-LGE-Proprietary
- */
-
-export 'primitive/font_size_primitive.dart';
-export 'primitive/font_weight_primitive.dart';
