--- lib/src/radius_tokens/radius_tokens.dart
@@ -1,7 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright 2025 LG Electronics Inc.
- * SPDX-License-Identifier: LicenseRef-LGE-Proprietary
- */
-
-export 'primitive/radius_primitive.dart';
-export 'semantic/radius_semantic.dart';
