--- lib/src/space_tokens/space_tokens.dart
@@ -1,6 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright 2025 LG Electronics Inc.
- * SPDX-License-Identifier: LicenseRef-LGE-Proprietary
- */
-
-export 'primitive/spacing_primitive.dart';
