+++ lib/src/webos_tokens/color_semantic/stroke/overlay/overlay.dart
