+++ lib/src/webos_tokens/color_semantic/stroke/stroke.dart
